refactor: :truck: rename price chart to historical chart
